Swift 4 ready
All files are Swift 4 for iOS 11. This is an independent update, as I needed to view these on iOS 11. I did the best I could with the time available.

Six filters in the CustomFilters.swift file are commented out in this version because they are not playing nice with Swift 4, and I have not had time to take a closer look at them.

The project is currently set for the iPad as that is what I use. It has not yet been tested on the iPhone. If it is needed for the iPhone, just change the project settings for your use.

In the ViewController.swift file there are five warnings about the deprecation of "topLayoutGuide." I haven't had time to fix that as I seldom use storyboards.

Issues:
1. Sometimes filters will crash when testing many in sequence. Just restart the app and try again, usually that fixes the problem.
2. When compiling and running from xCode to my iPad, the Metal filters may crash.   Finish running and then test the installed app, the Metal filters then work.